pyproject.toml
[tool.poetry]
name = "trello"
version = "0.1.0"
description = ""
authors = ["Dimitri Maksimov <dmitry-maksimov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.2.2"
selenium = "4.21.0"
allure-pytest = "2.13.5"
requests = "2.32.3"

[tool.pytest.ini_options]
addopts = [
    "--alluredir", "allure-results",
    "--clean-alluredir",
    "-s",
    "-v",
]

testpaths = [
    "tests/api",
    "tests/ui",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
